Patch for IP pool issues:
The problem is when we have more than one api-server, when a subnet is
deleted, the other api-servers keep the subnet allocator in their
subnet_allocator dict (in vnc_db.py). So when we create a subnet with
the same parameters (net, name, cidr,...) when we attach a port to him,
the allocated api is not always in the good range, depending on which
api-server responded to the create-port request.

- net_(delete/update)_notify methods modified,to update dicts on each nodes.

- ipam_(delete/update)_notify modified too.

- new parameter added (notify) now, "IndexAllocator.delete_all"
  is done only from "*_req" methods.

- unittests updated to match and test new behavior

Closes-Bug: #1332645
Closes-Bug: #1365824
(cherry picked from commit f3056c8663be341484468843888ce1646bd5da5d)

+++ src/config/api-server/vnc_db.py
@@ -458,6 +458,7 @@ class VncZkClient(object):
                                 addr_from_start, should_persist,
                                 start_subnet, size, alloc_unit):
         # TODO handle subnet resizing change, ignore for now
+
         if subnet not in self._subnet_allocators:
             if addr_from_start is None:
                 addr_from_start = False
@@ -470,10 +471,12 @@ class VncZkClient(object):
                 max_alloc=self._MAX_SUBNET_ADDR_ALLOC/alloc_unit)
     # end create_subnet_allocator
 
-    def delete_subnet_allocator(self, subnet):
-        self._subnet_allocators.pop(subnet, None)
-        IndexAllocator.delete_all(self._zk_client,
-                                  self._subnet_path+'/'+subnet+'/')
+    def delete_subnet_allocator(self, subnet, notify=True):
+        if subnet in self._subnet_allocators:
+            self._subnet_allocators.pop(subnet, None)
+        if not notify:
+            IndexAllocator.delete_all(self._zk_client,
+                        self._subnet_path+'/'+subnet+'/')
     # end delete_subnet_allocator
 
     def _get_subnet_allocator(self, subnet):
@@ -1299,8 +1302,8 @@ class VncDbClient(object):
             should_persist, start_subnet, size, alloc_unit)
     # end subnet_create_allocator
 
-    def subnet_delete_allocator(self, subnet):
-        return self._zk_db.delete_subnet_allocator(subnet)
+    def subnet_delete_allocator(self, subnet, notify=True):
+        return self._zk_db.delete_subnet_allocator(subnet, notify)
     # end subnet_delete_allocator
 
     def uuid_vnlist(self):
